feat(cmake): Enabling building both unit tests and main project with CMake

+++ unitTest.sh
@@ -1,12 +1,18 @@
 #!/bin/bash
-# Dependency porcess
+# Dependency process
 # Traversal of globalDependencies.json
 Vendors="Vendors"
 source $(pwd)/Shells/installVendor.sh && \
 dependenciesTraversal $(pwd)/Settings/.Json/globalDependencies.json $(pwd)/$Vendors/.$Vendors.json
 
+ORANGE='\033[0;33m'
+NC='\033[0m'
+
 # Cmake process
-rm -rf build
+if [[ "$1" = "--remove" ]]; then
+	rm -rf build
+	printf "${ORANGE}The folder, \"build\", directory has been removed.${NC}\n"
+fi
 mkdir -p build
 cmake -S . -B build
 cmake --build build
